requirements.txt
cryptography
dash
flask
flask-login
flask-mail
flask-sqlalchemy
gunicorn
plotly
pymysql
pandas
openai
alpha-vantage
python-dotenv  # Load .env variables
